refactor: move EmailSender to utils and add ImageUploader for Cloudinary integration
