+++ src/main/java/com/greenspace/api/utils/EmailSender.java
@@ -1,4 +1,4 @@
-package com.greenspace.api.features.email;
+package com.greenspace.api.utils;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
